+++ workspace_secretary/db/__init__.py
@@ -1,0 +1,8 @@
+from workspace_secretary.db.types import DatabaseConnection, DatabaseInterface
+from workspace_secretary.db.postgres import PostgresDatabase
+
+__all__ = [
+    "DatabaseConnection",
+    "DatabaseInterface",
+    "PostgresDatabase",
+]